Move some tests to new EnggUtils test file
