+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/calibration/test/test_calib_model.py
@@ -7,11 +7,10 @@
 import unittest
 from unittest import mock
 from unittest.mock import call, patch, create_autospec
-from numpy import array
 from os import path
-from Engineering.EnggUtils import GROUP, read_diff_constants_from_prm, create_output_files
+from Engineering.EnggUtils import GROUP
 from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.calibration.model import CalibrationModel
-from Engineering.common.calibration_info  import CalibrationInfo
+from Engineering.common.calibration_info import CalibrationInfo
 from testhelpers import assert_any_call_partial
 
 file_path = "mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.calibration.model"
@@ -35,69 +34,6 @@ class CalibrationModelTest(unittest.TestCase):
         mock_load_full_calib.assert_called_once()
         mock_write_diff_consts.assert_called_once_with(self.calibration_info.prm_filepath)
         self.calibration_info.load_relevant_calibration_files.assert_called_once()
-
-    def test_read_diff_constants_from_prm(self):
-        file_content = """ID    ENGIN-X CALIBRATION WITH CeO2 and V-Nb
-INS    CALIB   241391   ceo2
-INS  1 ICONS  18306.98      2.99     14.44
-INS  2 ICONS  18497.75    -29.68    -26.50"""
-        mocked_handle = mock.mock_open(read_data=file_content)
-        dummy_file_path = "/foo/bar_123.prm"
-        patchable = "builtins.open"
-        with mock.patch(patchable, mocked_handle):
-            diff_consts = read_diff_constants_from_prm(dummy_file_path)
-        deltas = abs(diff_consts - array([[2.99, 18306.98, 14.44], [-29.68, 18497.75, -26.5]]))
-        self.assertTrue((deltas < 1e-10).all())
-
-    @patch(enggutils_path + ".copy2")
-    @patch(enggutils_path + ".makedirs")
-    @patch(enggutils_path + ".path.exists")
-    @patch(enggutils_path + ".mantid.SaveNexus")
-    @patch(enggutils_path + ".write_prm_file")
-    def test_create_output_files_makes_savdir_and_saves_both_banks(self, mock_write_prm, mock_save_nxs, mock_exists,
-                                                                   mock_mkdir, mock_copy):
-        mock_exists.return_value = False  # make new directory
-        calibration = CalibrationInfo()  # easier to work with real calibration info object here
-        prm_name = "ENGINX_193749_all_banks.prm"
-        calibration.set_calibration_from_prm_fname(prm_name)
-        calibration.set_calibration_table("cal_table")
-        save_dir = "savedir"
-
-        create_output_files(save_dir, calibration, "ws")
-
-        mock_mkdir.assert_called_once_with(save_dir)
-        self.calibration_info.save_grouping_workspace.assert_not_called()  # only called if not bank data
-        prm_fpath = path.join(save_dir, prm_name)
-        write_prm_calls = [call("ws", prm_fpath),
-                           call("ws", prm_fpath.replace("all_banks", "bank_1"), spec_nums=[0]),
-                           call("ws", prm_fpath.replace("all_banks", "bank_2"), spec_nums=[1])]
-        mock_write_prm.assert_has_calls(write_prm_calls)
-        nxs_fpath = prm_fpath.replace(".prm", ".nxs")
-        mock_save_nxs.assert_called_once_with(InputWorkspace="cal_table", Filename=nxs_fpath)
-        copy_calls = [call(nxs_fpath, nxs_fpath.replace("all_banks", "bank_1")),
-                      call(nxs_fpath, nxs_fpath.replace("all_banks", "bank_2"))]
-        mock_copy.assert_has_calls(copy_calls)
-
-    @patch(enggutils_path + ".copy2")
-    @patch(enggutils_path + ".path")
-    @patch(enggutils_path + ".mantid.SaveNexus")
-    @patch(enggutils_path + ".write_prm_file")
-    def test_create_output_files_saves_custom_group_file(self, mock_write_prm, mock_save_nxs, mock_path, mock_copy):
-        mock_path.exists.return_value = True
-        prm_fname = "prm.prm"
-        mock_path.join.return_value = prm_fname
-        mock_path.splitext.return_value = (prm_fname.replace(".prm", ""), None)
-        self.calibration_info.group.banks = None  # no bank data e.g. custom
-        self.calibration_info.generate_output_file_name.return_value = prm_fname
-        self.calibration_info.get_calibration_table.return_value = "cal_table"  # no bank data e.g. custom
-        save_dir = "savedir"
-
-        create_output_files(save_dir, self.calibration_info, "ws")
-
-        self.calibration_info.save_grouping_workspace.assert_called_once_with(save_dir)
-        mock_write_prm.assert_called_once_with("ws", prm_fname)
-        mock_save_nxs.assert_called_once_with(InputWorkspace="cal_table", Filename=prm_fname.replace(".prm", ".nxs"))
-        mock_copy.assert_not_called()
 
     @patch(enggutils_path + '.mantid.DeleteWorkspace')
     @patch(enggutils_path + '.create_output_files')
