+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/focus/test/test_focus_model.py
@@ -10,7 +10,7 @@ import shutil
 from os import path
 
 from unittest.mock import patch, MagicMock, call, create_autospec
-from Engineering.EnggUtils import GROUP, _save_output_files, _load_run_and_convert_to_dSpacing, process_vanadium
+from Engineering.EnggUtils import GROUP
 from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.focus import model
 from Engineering.common.calibration_info import CalibrationInfo
 
@@ -143,135 +143,6 @@ class FocusModelTest(unittest.TestCase):
         mock_plot.assert_not_called()
         mock_save_out.assert_not_called()
         mock_del_ws.assert_not_called()
-
-    @patch(enggutils_path + '.path_handling.get_run_number_from_path')
-    @patch(enggutils_path + '.ADS')
-    def test_process_vanadium_foc_curves_exist(self, mock_ads, mock_path):
-        mock_path.return_value = "123456"
-        mock_ads.doesExist.return_value = True  # foc vanadium exist
-        mock_ads.retrieve.return_value = "van_ws_foc"
-
-        ws_van_foc, van_run = process_vanadium("van_path", self.calibration, "full_calib")
-
-        self.assertEqual(ws_van_foc, "van_ws_foc")
-        self.assertEqual(van_run, "123456")
-
-    @patch(enggutils_path + '._smooth_vanadium')
-    @patch(enggutils_path + '._focus_run_and_apply_roi_calibration')
-    @patch(enggutils_path + '._load_run_and_convert_to_dSpacing')
-    @patch(enggutils_path + '.path_handling.get_run_number_from_path')
-    @patch(enggutils_path + '.ADS')
-    def test_process_vanadium_run_exists_not_focused_over_ROI(self, mock_ads, mock_path, mock_load_run,
-                                                              mock_foc_run, mock_smooth_van):
-        mock_path.return_value = "123456"
-        mock_ads.doesExist.side_effect = [False, True]  # foc vanadium not exist but original van ws does
-        mock_smooth_van.return_value = "van_ws_foc"  # last alg called before return
-
-        ws_van_foc, van_run = process_vanadium("van_path", self.calibration, "full_calib")
-
-        mock_ads.retrieve.assert_called_once_with("123456")
-        mock_load_run.assert_not_called()
-        mock_foc_run.assert_called_once()
-        mock_smooth_van.assert_called_once()
-        self.assertEqual(ws_van_foc, "van_ws_foc")
-        self.assertEqual(van_run, "123456")
-
-    @patch(enggutils_path + '._smooth_vanadium')
-    @patch(enggutils_path + '._focus_run_and_apply_roi_calibration')
-    @patch(enggutils_path + '._load_run_and_convert_to_dSpacing')
-    @patch(enggutils_path + '.path_handling.get_run_number_from_path')
-    @patch(enggutils_path + '.ADS')
-    def test_process_vanadium_run_not_loaded(self, mock_ads, mock_path, mock_load_run, mock_foc_run, mock_smooth_van):
-        mock_path.return_value = "123456"
-        mock_ads.doesExist.side_effect = [False, False]  # vanadium run not loaded
-        mock_smooth_van.return_value = "van_ws_foc"  # last alg called before return
-
-        ws_van_foc, van_run = process_vanadium("van_path", self.calibration, "full_calib")
-
-        mock_ads.retrieve.assert_not_called()
-        mock_load_run.assert_called_once()
-        mock_foc_run.assert_called_once()
-        mock_smooth_van.assert_called_once()
-        self.assertEqual(ws_van_foc, "van_ws_foc")
-        self.assertEqual(van_run, "123456")
-
-    @patch(enggutils_path + '.mantid.DeleteWorkspace')
-    @patch(enggutils_path + '.mantid.NormaliseByCurrent')
-    @patch(enggutils_path + '.logger')
-    @patch(enggutils_path + '.path_handling.get_run_number_from_path')
-    @patch(enggutils_path + '.mantid.Load')
-    def test_load_runs_ignores_empty_runs_with_zeros_charge(self, mock_load, mock_path, mock_log, mock_norm, mock_del):
-        ws = MagicMock()
-        ws.getRun.return_value = MagicMock()
-        ws.getRun().getProtonCharge.return_value = 0  # zero proton charge -> empty run to be ignored
-        mock_load.return_value = ws
-
-        ws_foc = _load_run_and_convert_to_dSpacing("fpath", "instrument", "full_calib")
-
-        self.assertIsNone(ws_foc)
-        mock_log.warning.assert_called_once()
-        mock_norm.assert_not_called()  # throws error if zero charge
-        mock_del.assert_called_once()
-
-    @patch(enggutils_path + '.path.exists')
-    @patch(enggutils_path + ".mantid.SaveFocusedXYE")
-    @patch(enggutils_path + ".mantid.SaveGSS")
-    @patch(enggutils_path + ".mantid.SaveNexus")
-    @patch(enggutils_path + ".mantid.AddSampleLog")
-    def test_save_output_files_both_banks_no_RB_number_path_exists(self, mock_add_log, mock_save_nxs, mock_save_gss,
-                                                                   mock_save_xye, mock_path):
-        mock_path.return_value = True  # directory exists
-        ws_foc = MagicMock()
-        ws_foc.getNumberHistograms.return_value = 2
-        ws_foc.getDimension.return_value = MagicMock()
-        ws_foc.getDimension().name = 'Time-of-flight'  # x-unit
-        ws_foc.run.return_value = MagicMock()
-        ws_foc.run().get.return_value = MagicMock().value
-        ws_foc.run().get().value = '193749'  # runno
-        van_run = '123456'
-
-        focused_files = _save_output_files(["save_dir"], ws_foc, self.calibration, van_run, rb_num=None)
-
-        mock_save_gss.assert_called_once()
-        mock_save_xye.assert_called_once()
-        add_log_calls = [call(Workspace=ws_foc, LogName="Vanadium Run", LogText=van_run),
-                         call(Workspace=ws_foc, LogName="bankid", LogText='bank 1'),
-                         call(Workspace=ws_foc, LogName="bankid", LogText='bank 2')]
-        mock_add_log.assert_has_calls(add_log_calls)
-        save_nxs_calls = [call(InputWorkspace=ws_foc, Filename=focused_files[0], WorkspaceIndexList=[0]),
-                          call(InputWorkspace=ws_foc, Filename=focused_files[1], WorkspaceIndexList=[1])]
-        mock_save_nxs.assert_has_calls(save_nxs_calls)
-
-    @patch(enggutils_path + '.makedirs')
-    @patch(enggutils_path + '.path.exists')
-    @patch(enggutils_path + ".mantid.SaveFocusedXYE")
-    @patch(enggutils_path + ".mantid.SaveGSS")
-    @patch(enggutils_path + ".mantid.SaveNexus")
-    @patch(enggutils_path + ".mantid.AddSampleLog")
-    def test_save_output_files_North_Bank_RB_number_path_not_exists(self, mock_add_log, mock_save_nxs, mock_save_gss,
-                                                                    mock_save_xye, mock_path, mock_mkdir):
-        self.calibration.get_group_suffix.return_value = "bank_1"
-        self.calibration.get_foc_ws_suffix.return_value = "bank_1"
-        mock_path.return_value = False  # directory exists
-        ws_foc = MagicMock()
-        ws_foc.getNumberHistograms.return_value = 1
-        ws_foc.getDimension.return_value = MagicMock()
-        ws_foc.getDimension().name = 'Time-of-flight'  # x-unit
-        ws_foc.run.return_value = MagicMock()
-        ws_foc.run().get.return_value = MagicMock().value
-        ws_foc.run().get().value = '193749'  # runno
-        van_run = '123456'
-        rb_num = '1'
-
-        focused_files = _save_output_files(["save_dir"], ws_foc, self.calibration, van_run, rb_num=rb_num)
-
-        mock_mkdir.assert_called_once()
-        mock_save_gss.assert_called_once()
-        mock_save_xye.assert_called_once()
-        add_log_calls = [call(Workspace=ws_foc, LogName="Vanadium Run", LogText=van_run),
-                         call(Workspace=ws_foc, LogName="bankid", LogText='bank 1')]
-        mock_add_log.assert_has_calls(add_log_calls)
-        mock_save_nxs.assert_called_once_with(InputWorkspace=ws_foc, Filename=focused_files[0], WorkspaceIndexList=[0])
 
 
 if __name__ == '__main__':
